IMPALA-6554, IMPALA-6565: Fix some bugs in DequeueDeferredRpc()
This change fixes 2 bugs in KrpcDataStreamRecvr::SenderQueue::DequeueDeferredRpc():

IMPALA-6554: Previously, RespondAndReleaseRpc() was called after enqueuing
a deferred batch into the row batch queue without holding the per sender-queue's
lock. This may end up racing with KrpcDataStreamRecvr::Close() and causes
MemTracker::Release() to be called after MemTracker::Close() has been called.
This change fixes the problem by making sure the lock is still held when
RespondAndReleaseRpc() is called.

IMPALA-6565: Previously, the local variable 'status' is unintentionally
defined twice in DequeueDeferredRpc(): one in the outer scope and one
in the inner scope. This causes the error status of AddBatchWork() to
be dropped when the inner scope ends. As a result, the error status from
AddBatchWork() (e.g. MemLimitExceeded) will not be propagated back to the
sender aand the receiver will continue to operate with some missing data,
leading to wrong query results. This change fixes the problem by removing
the inner scope and avoids the re-definition of the status local variable.

Testing done: stress test.

Change-Id: I6b2985a47021ebd4a970861040e7474aca7941b5